requirements.txt
opencv-python
opencv-contrib-python
face_recognition
numpy
pandas
Pillow
Tk

